Modify tests to meet the project requirements
